[NCL-8071] Create endpoint for listing Deliverable artifacts

+++ model/src/main/resources/db-update/20070-from-2.6.0-to-2.7.0.sql
@@ -87,7 +87,7 @@ CREATE TABLE deliverableartifact
     report_id           BIGINT                  NOT NULL,
     artifact_id         INTEGER                 NOT NULL,
     builtfromsource     BOOLEAN                 NOT NULL,
-    brewbuildid         INTEGER,
+    brewbuildid         BIGINT,
     PRIMARY KEY (report_id, artifact_id)
 );
 ALTER TABLE deliverableartifact
